trying to fix CI test errors for ubuntu

+++ source/config/config.h
@@ -66,13 +66,13 @@ class Option
             }
             throw ConfigError::cast_error<std::string>(_name, _value);
         } else if constexpr (std::is_integral_v<T> && !std::is_same_v<T, bool>) {
-            if (is<long long>()) {
-                auto value = boost::any_cast<long long>(_value);
+            if (is<int64_t>()) {
+                auto value = boost::any_cast<int64_t>(_value);
                 return static_cast<T>(value);
             } else if (_value.empty()) {
                 return default_value;
             }
-            throw ConfigError::cast_error<long long>(_name, _value);
+            throw ConfigError::cast_error<int64_t>(_name, _value);
         } else {
             if (is<T>()) {
                 return boost::any_cast<T>(_value);
